# delivery-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Management System API
  version: 1.0.0
  description: |
    OpenAPI spec for the Delivery Management System backend.
    Note: Admin accounts are intentionally not exposed in the frontend UI; use manual DB insertion or an admin-only backend procedure to create admin users.
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, seller, admin]
    AuthRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role:
          type: string
          enum: [buyer, seller, admin]
    Order:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        description: { type: string }
        buyerId:
          { oneOf: [{ type: string }, { $ref: "#/components/schemas/User" }] }
        sellerId:
          { oneOf: [{ type: string }, { $ref: "#/components/schemas/User" }] }
        status: { type: string }
        history:
          type: array
          items:
            type: object
            properties:
              stage: { type: string }
              at: { type: string, format: date-time }
paths:
  /auth/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered successfully (returns token and user)
        "400":
          description: Validation error or bad request
      description: |
        Creates a user. NOTE: frontend hides `admin` role; creating admin should be done manually or with appropriate safeguards.
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Returns JWT and user object
  /orders:
    post:
      summary: Create an order (Buyer)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        "201": { description: Created }
    get:
      summary: List orders (role-dependent)
      security:
        - bearerAuth: []
      responses:
        "200": { description: OK }
  /orders/{id}/assign-seller:
    put:
      summary: Assign a seller to an order (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sellerId: { type: string }
      responses:
        "200": { description: Seller assigned }
        "403": { description: Forbidden (admin-only) }
  /orders/{id}/next-stage:
    put:
      summary: Advance order stage (Seller only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Stage advanced }
  /orders/{id}:
    get:
      summary: Get order details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order object
  /stats/all:
    get:
      summary: Admin stats (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200": { description: Stats object }
security:
  - bearerAuth: []